linked up delte buttons to delete 100 news and 100 posts from the database

+++ WNGA/src/main/webapp/jsp/NewsFeed.jsp
@@ -75,7 +75,7 @@
                                 <sec:authorize access="hasRole('ROLE_ADMIN')">
                                     <p class="edit-delete">
                                         <a class="grey" href="editNewsForm?newsFeedId=${news.newsFeedId}">edit</a> 
-                                        <a class="grey" href="deleteNews?newsFeedId=${news.newsFeedId}">delete</a>
+                                        <a class="grey" href="deleteNews?newsFeedId=${news.newsFeedId} id="delete" data-confirm=" WARNING!!!!: Are you sure to delete this comment? CANNOT BE UNDONE!!!"">delete</a>
                                     </p>
                                 </sec:authorize>
                             </div>
